Reorganizing the files. ipython notebook and user_test_submission comes to main (root). We add data/public for public data sets, both notebook and user test reads from there. We add submissions/<submission_name> for example submissions. We will require at least submissions/starting_kit/*.py which user_test_submission will execute, but organizer can provide other example submissions, too.
