+++ .travis.yml
@@ -11,8 +11,8 @@ install:
     - conda install --yes --quiet numpy scipy matplotlib scikit-learn pandas
     - pip install -q flake8
 script:
-    - cd starting_kit
-    - flake8 *.py
+    - flake8 user_test_submission.py
+    - flake8 submissions/*/*.py
     - python user_test_submission.py
 notifications:
   email: true
